Add unit tests for validation methods in VariantContext
There will be unit tests that test the functionality of:
extraStrictValidation
validateReferenceBases
validateRSIDs(rsIDs);
validateAlternateAlleles();
validateChromosomeCounts();

The extraStrictValidation method calls all the other methods,
so the test for this makes use of the data providers for the
other tests - I couldn't think of a better way to do this,
and someone could change what extraStrictValidation does, so
it seemed safest to provide a test for it.  Each method has
a pass and a fail test for all the conditions I could think of.

Resolves: #203